test(logic): split shared-validator suite under 400-line gate (Refs #2394)
`packages/colonizethis_logic/test/orders/order_suggestion_shared_validator_test.dart`
was 492 physical lines and tripped the `repo.logic_test_file_size` gate
(<= 400 physical lines per file). Split it by concern, mirroring the
existing `incremental_candidate_validator_equivalence_test_helpers.dart`
pattern of co-locating fixtures next to the suites that consume them:

- `..._test_helpers.dart` (126 lines): province ID constants, `buildGame`,
  `buildTopology` corpus shared by both suites.
- `..._equivalence_test.dart` (217 lines): five `suggest*Orders` shared-
  validator equivalence cases plus the pre-built `view`/`unitsById`
  forPlayer equivalence case.
- `..._negative_test.dart` (173 lines): five mismatched-playerId
  assertion cases (one per suggest family + `forPlayer` foreign view)
  and the `generateOrdersWithSimpleHeuristics` smoke test.

No behavior change; same test bodies, same expectations. Verified:
- `dart analyze test/orders/order_suggestion_shared_validator_*.dart` clean
- `dart test test/orders/order_suggestion_shared_validator_*.dart` -> +11 All tests passed
- `dart run tool/check_logic_test_file_size.dart <new files>` -> no violations
